feat(resources): fix eslint bugs
